認証入力部分をコンポーネント化
